+++ examples/demo.rs
@@ -61,8 +61,7 @@ fn main() {
         let mut layout = elements::LinearLayout::vertical();
         if page > 1 {
             layout.push(
-                elements::Paragraph::new(format!("Page {}", page))
-                    .aligned(Alignment::Center),
+                elements::Paragraph::new(format!("Page {}", page)).aligned(Alignment::Center),
             );
             layout.push(elements::Break::new(1));
         }
@@ -184,9 +183,7 @@ fn main() {
     doc.push(elements::Paragraph::new(
         "You already saw lists and formatted centered text. Here are some other examples:",
     ));
-    doc.push(
-        elements::Paragraph::new("This is right-aligned text.").aligned(Alignment::Right),
-    );
+    doc.push(elements::Paragraph::new("This is right-aligned text.").aligned(Alignment::Right));
     doc.push(
         elements::Paragraph::new("And this paragraph has a frame drawn around it and is colored.")
             .padded(genpdf::Margins::vh(0, 1))
@@ -207,7 +204,9 @@ fn main() {
     );
     doc.push(elements::Break::new(1.5));
 
-    doc.push(elements::Paragraph::new("Embedding images also works using the 'images' feature."));
+    doc.push(elements::Paragraph::new(
+        "Embedding images also works using the 'images' feature.",
+    ));
     #[cfg(feature = "images")]
     images::do_image_test(&mut doc);
 
@@ -309,16 +308,23 @@ mod images {
     const IMAGE_PATH_JPG: &'static str = "examples/images/test_image.jpg";
 
     pub fn do_image_test(doc: &mut genpdf::Document) {
-        doc.push(elements::Paragraph::new("Here is an example image with default position/scale:"));
+        doc.push(elements::Paragraph::new(
+            "Here is an example image with default position/scale:",
+        ));
         doc.push(elements::Image::from_path(IMAGE_PATH_JPG).expect("Unable to load image"));
-        doc.push(elements::Paragraph::new("and here is one that is centered, rotated, and scaled some."));
+        doc.push(elements::Paragraph::new(
+            "and here is one that is centered, rotated, and scaled some.",
+        ));
         doc.push(
             elements::Image::from_path(IMAGE_PATH_JPG)
                 .expect("Unable to load image")
                 .with_alignment(Alignment::Center)
                 .with_scale(genpdf::Scale::new(0.5, 2))
-                .with_clockwise_rotation(45.0)
+                .with_clockwise_rotation(45.0),
         );
-        doc.push(elements::Paragraph::new("For a full example of image functionality, please see images.pdf."));
+        doc.push(elements::Paragraph::new(
+            "For a full example of image functionality, please see images.pdf.",
+        ));
+        doc.push(elements::Break::new(1.5));
     }
 }