+++ examples/images.rs
@@ -1,10 +1,16 @@
+// SPDX-FileCopyrightText: 2020 Alexander Dean-Kennedy <dstar@slackless.com>
+// SPDX-License-Identifier: CC0-1.0
+
 use std::env;
 
 use genpdf::Alignment;
 use genpdf::Element as _;
 use genpdf::{elements, fonts, style};
 
-const FONT_DIR: &'static str = "/usr/share/fonts/truetype/liberation";
+const FONT_DIRS: &[&str] = &[
+    "/usr/share/fonts/liberation",
+    "/usr/share/fonts/truetype/liberation",
+];
 const DEFAULT_FONT_NAME: &'static str = "LiberationSans";
 
 const IMAGE_PATH_JPG: &'static str = "examples/images/test_image.jpg";
@@ -18,8 +24,13 @@ fn main() {
     }
     let output_file = &args[0];
 
+    let font_dir = FONT_DIRS
+        .iter()
+        .filter(|path| std::path::Path::new(path).exists())
+        .next()
+        .expect("Could not find font directory");
     let default_font =
-        fonts::from_files(FONT_DIR, DEFAULT_FONT_NAME, Some(fonts::Builtin::Helvetica))
+        fonts::from_files(font_dir, DEFAULT_FONT_NAME, Some(fonts::Builtin::Helvetica))
             .expect("Failed to load the default font family");
 
     let mut doc = genpdf::Document::new(default_font);
@@ -33,8 +44,7 @@ fn main() {
         let mut layout = elements::LinearLayout::vertical();
         if page > 1 {
             layout.push(
-                elements::Paragraph::new(format!("Page {}", page))
-                    .aligned(Alignment::Center),
+                elements::Paragraph::new(format!("Page {}", page)).aligned(Alignment::Center),
             );
             layout.push(elements::Break::new(1));
         }
@@ -56,7 +66,7 @@ fn main() {
         elements::Image::from_path(IMAGE_PATH_JPG)
             .expect("Unable to load alt image")
             .with_position(genpdf::Position::new(170, -10)) // far over to right and down
-            .with_clockwise_rotation(90.0)
+            .with_clockwise_rotation(90.0),
     );
 
     // adding a break to avoid the image posted above with an "absolute image.
@@ -64,7 +74,7 @@ fn main() {
 
     // IMAGE FILE TYPE HANDLING:
     doc.push(elements::Paragraph::new(
-        "Table with image format/scaling tests:"
+        "Table with image format/scaling tests:",
     ));
     let mut img_table = elements::TableLayout::new(vec![2, 2, 2, 2]);
     img_table.set_cell_decorator(elements::FrameCellDecorator::new(true, true, false));
@@ -75,9 +85,11 @@ fn main() {
         .element(elements::Text::new("Half Size").padded(1))
         .element(elements::Text::new("Double Size").padded(1))
         .push()
-        .expect("Invalid Row.");
+        .expect("Invalid row");
     for (ftype, path) in vec![
-        ("BMP", IMAGE_PATH_BMP), ("JPG", IMAGE_PATH_JPG), ("PNG", IMAGE_PATH_PNG)
+        ("BMP", IMAGE_PATH_BMP),
+        ("JPG", IMAGE_PATH_JPG),
+        ("PNG", IMAGE_PATH_PNG),
     ] {
         let img = elements::Image::from_path(path).expect("invalid image");
         img_table
@@ -87,13 +99,13 @@ fn main() {
             .element(img.clone().with_scale(genpdf::Scale::new(0.5, 0.5)))
             .element(img.clone().with_scale(genpdf::Scale::new(2, 2)))
             .push()
-            .expect("Invalid row.");
+            .expect("Invalid row");
     }
     doc.push(img_table);
 
     doc.push(elements::Break::new(2));
     doc.push(elements::Paragraph::new(
-        "Table with image rotation/offset calculation tests:"
+        "Table with image rotation/offset calculation tests:",
     ));
     let mut rot_table = elements::TableLayout::new(vec![2, 2, 2, 2, 2, 2, 2]);
     rot_table.set_cell_decorator(elements::FrameCellDecorator::new(true, true, false));
@@ -102,12 +114,12 @@ fn main() {
         .element(elements::Text::new("Rot").padded(1))
         .element(elements::Text::new("30°").padded(1))
         .element(elements::Text::new("45°").padded(1))
-        .element(elements::Text::new("90").padded(1))
-        .element(elements::Text::new("120").padded(1))
-        .element(elements::Text::new("150").padded(1))
-        .element(elements::Text::new("180").padded(1))
+        .element(elements::Text::new("90°").padded(1))
+        .element(elements::Text::new("120°").padded(1))
+        .element(elements::Text::new("150°").padded(1))
+        .element(elements::Text::new("180°").padded(1))
         .push()
-        .expect("Invalid Row.");
+        .expect("Invalid row");
     let img = elements::Image::from_path(IMAGE_PATH_JPG).expect("invalid image");
     rot_table
         .row()
@@ -119,7 +131,7 @@ fn main() {
         .element(img.clone().with_clockwise_rotation(150.0))
         .element(img.clone().with_clockwise_rotation(180.0))
         .push()
-        .expect("Invalid Row.");
+        .expect("Invalid row");
     rot_table
         .row()
         .element(elements::Text::new("Negative").padded(1))
@@ -130,9 +142,8 @@ fn main() {
         .element(img.clone().with_clockwise_rotation(-150.0))
         .element(img.clone().with_clockwise_rotation(-180.0))
         .push()
-        .expect("Invalid Row.");
+        .expect("Invalid row");
     doc.push(rot_table);
-
 
     doc.render_to_file(output_file)
         .expect("Failed to write output file");